\name{arity}
\alias{arity}
\title{Determine the number of arguments of a function...}
\usage{arity(f)
}
\description{Determine the number of arguments of a function}
\details{Tries to determine the number of arguments of function.}
\value{The arity of the function \code{f}.}
\arguments{\item{f}{The function to determine the arity for.}
}

