% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster_size}
\alias{cluster_size}
\title{Pathway sizes in terms of diseases and drugs according to KEGG}
\format{
A numeric vector with 304 elements:
}
\source{
\url{https://www.kegg.jp/}
}
\usage{
cluster_size
}
\description{
A dataset the numbers of drugs and diseases associated with
human pathways from KEGG.
}
\keyword{datasets}
