% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_create_cohort.R
\name{create_cohort}
\alias{create_cohort}
\title{Simulate Case-Control Data with Correlated Predictors}
\usage{
create_cohort(
  nobs,
  npred,
  ccr,
  nassoc,
  marginal_prob,
  true_beta,
  corr_param = NULL,
  bal = FALSE
)
}
\arguments{
\item{nobs}{Integer, number of observations.}

\item{npred}{Integer, number of predictor variables.}

\item{ccr}{Numeric, case-control ratio}

\item{nassoc}{Integer, number of predictors that are truely
associated with the outcome.}

\item{marginal_prob}{A numeric vector of size \code{npred} that determines the
marginal probabilities of the predictor variables.}

\item{true_beta}{numeric vector of true effects.}

\item{corr_param}{either a correlation structure between predictor variables as numeric.
matrix or rho as numeric for \code{genCorGen()} (Default: NULL).}

\item{bal}{if TRUE, the case-to-control ratio is printed (Default: FALSE).}
}
\value{
A data.table object of size \code{nobs} x (\code{npred} + 2).
}
\description{
This function uses \code{simstudy} package and creates a dataset of
a cohort based on given parameter.
}
\examples{
library(simstudy)

set.seed(4875)
nobs <- 1000
npred <- 100
ccr <- 0.5
nassoc <- 5
marginal_prob <- rep(0.15, npred)
true_beta <- rep(c(log(5), 0), c(nassoc, npred - nassoc))

# -- simulate data
cohort <- create_cohort(nobs = nobs, npred = npred, nassoc = nassoc, ccr = ccr,
                        marginal_prob = marginal_prob, true_beta = true_beta)

# -- test marginals and effects
apply(cohort, 2, mean)
summary(glm(out ~ ., data = cohort[,-1], family = "binomial", maxit = 100))$coefficients


# -- simulate correlated data (positive-definite)
corr_param <- outer(1:npred, 1:npred,
                     function(x,y) {(-1)^(1+abs(x-y)) * 0.95^abs(x-y)})

# -- or provide rho
corr_param <- 0.1

cohort <- create_cohort(nobs = nobs, npred = npred, nassoc = nassoc,
                        ccr = ccr, marginal_prob = marginal_prob,
                        true_beta = true_beta, corr_param = corr_param)

}
\seealso{
\code{\link[simstudy]{genCorGen}}.
}
