% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftarget_db_manager.R
\name{drug2atc2target}
\alias{drug2atc2target}
\title{find drug target: link drug to target ("pathway" or gene: "hsa") + trim first 4 characters (atc:)}
\usage{
drug2atc2target(drug_atc_target_file, kegg_target)
}
\description{
find drug target: link drug to target ("pathway" or gene: "hsa") + trim first 4 characters (atc:)
}
