% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftarget_db_manager.R
\name{ds2icd2target}
\alias{ds2icd2target}
\title{Find disease target: link disease to target}
\usage{
ds2icd2target(ds_icd_target_file, kegg_target)
}
\description{
Find disease target: link disease to target
}
