\name{subexpressions}
\alias{subexpressions}
\title{Functions for decomposing and recombining R expressions}
\usage{
  subexpressions(expr)
}
\arguments{
  \item{expr}{An R expression.}
}
\value{
  The decomposed or recombined expression.
}
\description{
  \code{subexpressions} returns a list of all
  subexpressions (subtrees) of an expression \code{expr}.
}

