% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{find_intercept_iter}
\alias{find_intercept_iter}
\title{Find the Intercept for a Dataset}
\usage{
find_intercept_iter(true_beta, dtx_old, ccr = 0.5, ...)
}
\arguments{
\item{true_beta}{A vector of true beta values including both assciated and
non-associated variables.}

\item{dtx_old}{Data.table of old dataset to be appended with outcome.}

\item{ccr}{Numeric, case-control ratio (Default: 0.5).}

\item{cepts}{A numeric list of intercepts to use, recommmended to be NULL
(Default: NULL).}

\item{iter}{Integer, number of increments between intercept limits (Default: 500).}
}
\value{
A list of
\itemize{
 \item \code{data}: The new data.table with outcome column \code{data$out}.
 \item \code{intercept}: An estimation of the intercept.
 \item \code{difference}: The absolute difference between \code{ccr}
and \code{mean(data$out)}.
 \item \code{sequence}: Sequence of tested intercepts.
 \item \code{ccr}: Case-to-control ratio for each intercept.
}
}
\description{
This function approximates intercept (\eqn{\beta_{0}})
iteratively to get a balanced case-control ratio in the simulated data.
This function uses \code{simstudy} package.
}
\examples{
dtx <- simstudy::genCorGen(n = 100, nvars = 5, params1 = 0.15,
 dist = "binary", wide = TRUE, rho = 0.1, corstr = "ar1")
find_intercept_iter(c(-2, -2, 2, 2, 2), dtx)

}
\seealso{
\code{\link[simstudy:defDataAdd]{defDataAdd}} and
\code{\link[simstudy:addColumns]{addColumns}} in \code{simstudy} package.
}
