% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inputVariablesOfIndividual}
\alias{inputVariablesOfIndividual}
\title{Functions for analysing GP individuals}
\usage{
inputVariablesOfIndividual(ind, inset)
}
\arguments{
\item{ind}{A GP individual, represented as a R function.}

\item{inset}{A set of input variables.}
}
\description{
\code{inputVariablesOfIndividual} returns a list of input variables in \code{inset} that are
used by the GP individual \code{ind}.
}

