% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.sType}
\alias{is.sType}
\title{Check if an object is an sType}
\usage{
is.sType(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
\code{TRUE} iff \code{x} is an sType.
}
\description{
Returns \code{TRUE} iff its argument is an sType.
}

