% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftarget_db_manager.R
\name{kegg_merge}
\alias{kegg_merge}
\title{Merge two data tables with the same key, usually kegg id}
\usage{
kegg_merge(kegg_table_1, kegg_table_2, key_column, merged_file_name)
}
\description{
Merge two data tables with the same key, usually kegg id
}
