% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mse}
\alias{mse}
\title{Mean squared error (MSE)}
\usage{
mse(x, y)
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list.}
}
\value{
The MSE between \code{x} and \code{y}.
}
\description{
Mean squared error (MSE)
}

