\name{new.alist}
\alias{new.alist}
\title{Create a new function argument list from a list or vector of strings}
\usage{
new.alist(fargs)
}
\arguments{
  \item{fargs}{The formal arguments, given as a list or
  vector of strings.}
}
\value{
A formal argument list, ready to be passed via
\code{\link{formals}}.
}
\description{
Creates a formal argument list from a list or vector of
strings, ready to be assigned via \code{\link{formals}}.
}

