\name{normalize}
\alias{normalize}
\title{Normalize a vector into the interval [0, 1]...}
\usage{normalize(x)
}
\description{Normalize a vector into the interval [0, 1]}
\value{The normalized vector.}
\arguments{\item{x}{The vector to normalize, so that each element lies in the
interval [0, 1].}
}

