\name{print.sType}
\alias{print.sType}
\title{Prints a sType and returns it invisible.}
\usage{
\method{print}{sType}(x, ...)
}
\arguments{
  \item{x}{The sType to print.}

  \item{...}{Optional parameters to print are ignored in
  this method.}
}
\description{
Prints a sType and returns it invisible.
}

