\name{r_ssse}
\alias{r_ssse}
\title{R version of Scaled sum squared error (sSSE)}
\usage{
r_ssse(x, y)
}
\arguments{
  \item{x}{A numeric vector or list.}

  \item{y}{A numeric vector or list.}
}
\value{
The sSSE between \code{x} and \code{y}.
}
\description{
R version of Scaled sum squared error (sSSE)
}

