\name{randelt}
\alias{randelt}
\title{Choose a random element from a list or vector...}
\usage{randelt(x)
}
\description{Choose a random element from a list or vector}
\details{Returns a unformly random chosen element of the vector or list \code{x}.}
\value{A uniformly random element of \code{x}.}
\arguments{\item{x}{The vector or list to chose an element from.}
}

