\name{randomExpressionCreation}
\alias{randexprGrow}
\title{Creates an R expression by random growth...}
\usage{randexprGrow(funcset, inset, conset, maxdepth=16, constprob=0.5, subtreeprob=0.5,
    curdepth=1)
randexprFull(funcset, inset, conset, maxdepth=16, constprob=0.5)
}
\description{Creates an R expression by random growth}
\details{\code{randexprGrow}: Creates a random R expression by randomly growing its tree. In each step of growth,
with probability \code{subtreeprob}, an operator is chosen from the function set \code{funcset}.
The operands are then generated by recursive calls. If no subtree is generated, a constant will
be generated with probability \code{constprob}. If no constant is generated, an input variable
will be chosen randomly. The depth of the resulting expression trees can be bounded by the
\code{maxdepth} parameter.
\code{randexprFull} creates a random full expression tree of depth \code{maxdepth}. The algorithm
is the same as \code{randexprGrow}, with the exception that the probability of generating
a subtree is fixed to 1  until the desired tree depth \code{maxdepth} is reached.

}
\value{\code{randexprGrow}: A new R expression generated by random growth.

}
\arguments{\item{funcset}{The function set.}
\item{inset}{The set of input variables.}
\item{conset}{The set of constant factories.}
\item{maxdepth}{The maximum expression tree depth.}
\item{constprob}{The probability of generating a constant in a step of growth, if no subtree
is generated. If neither a subtree nor a constant is generated, a randomly chosen input variable
will be generated.}
\item{subtreeprob}{The probability of generating a subtree in a step of growth.}
\item{curdepth}{(internal) The depth of the random expression currently generated, used internally
in recursive calls.}
}
\alias{randexprFull}

