\name{rgp-package}
\alias{rgp-package}
\alias{rgp-package}
\title{The RGP package}
\description{RGP is a simple yet flexible Genetic Programming system for the R
environment. The system implements classical untyped tree-based genetic
programming as well as more advanced variants including, for example,
strongly typed genetic pro- gramming and Pareto genetic programming.}
\docType{package}
\author{Oliver Flasch \email{oliver.flasch@fh-koeln.de}, Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}, Joerg Stork}
\keyword{package}

