\name{rsquared}
\alias{rsquared}
\title{Coefficient of determination (R^2)}
\usage{
rsquared(x, y)
}
\arguments{
  \item{x}{A numeric vector or list.}

  \item{y}{A numeric vector or list.}
}
\value{
The coefficient of determination (R^2) between \code{x} and
\code{y}.
}
\description{
Coefficient of determination (R^2)
}

