\name{smse}
\alias{smse}
\title{Scaled mean squared error (SMSE)...}
\usage{smse(x, y)
}
\description{Scaled mean squared error (SMSE)}
\details{Calculates the MSE between vectors after normalizing them into the
interval [0, 1].}
\value{The SMSE between \code{x} and \code{y}.}
\arguments{\item{x}{A numeric vector or list.}
\item{y}{A numeric vector or list.}
}

