\name{sortByRange}
\alias{sortByRange}
\title{Tabulate a list of functions or input variables by their range sTypes...}
\usage{sortByRange(x)
}
\description{Tabulate a list of functions or input variables by their range sTypes}
\value{A table of the objects keyed by their range types.}
\arguments{\item{x}{A list of functions or input variables to sort by range sType.}
}

