% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortByRanking}
\alias{sortByRanking}
\title{Sort a vector or list via a given ranking}
\usage{
sortByRanking(xs, ranking = rank(xs))
}
\arguments{
\item{xs}{The vector or list to reorder.}

\item{ranking}{The ranking to sort \code{xs} by, defaults to \code{rank(xs)}.}
}
\value{
The result of reordering \code{xs} by \code{ranking}.
}
\description{
Reorders a vector or list according to a given ranking \code{ranking}.
}

