\name{sortByType}
\alias{sortByType}
\title{Tabulate a list of functions or input variables by their sTypes...}
\usage{sortByType(x)
}
\description{Tabulate a list of functions or input variables by their sTypes}
\value{A table of the objects keyed by their sTypes.}
\arguments{\item{x}{A list of functions or input variables to sort by sType.}
}

