% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftarget_db_manager.R
\name{source2target}
\alias{source2target}
\title{Fetch KEGG Entries using KEGGREST}
\usage{
source2target(kegg_source, kegg_target, file_name = NULL)
}
\description{
Fetch KEGG Entries using KEGGREST
}
