% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sse}
\alias{sse}
\title{Sum squared error (SSE)}
\usage{
sse(x, y)
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list.}
}
\value{
The SSE between \code{x} and \code{y}.
}
\description{
Sum squared error (SSE)
}

