% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ssse}
\alias{ssse}
\title{Scaled sum squared error (sSSE)}
\usage{
ssse(x, y)
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list.}
}
\value{
The sSSE between \code{x} and \code{y}.
}
\description{
Scaled sum squared error (sSSE)
}

