\name{withAttributesOf}
\alias{withAttributesOf}
\title{Return expr with the attributes of sourceExpr...}
\usage{withAttributesOf(expr, sourceExpr)
}
\description{Return expr with the attributes of sourceExpr}
\value{A copy of \code{expr}, tagged with the attributes and sType of
\code{sourceExpr}.}
\arguments{\item{expr}{The expression to receive the attributes of \code{sourceExpr}.}
\item{sourceExpr}{The expression to provide the attributes for \code{expr}.}
}

