% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildingBlockTag}
\alias{buildingBlockTag}
\alias{buildingBlockTag<-}
\alias{hasBuildingBlockTag}
\title{Building block tags}
\usage{
buildingBlockTag(x)

buildingBlockTag(x) <- value

hasBuildingBlockTag(x)
}
\arguments{
\item{x}{An expression node.}

\item{value}{The value of the building block tag. Must be a numerical
  in the interval [0.0 1.0].}
}
\description{
To implement buidling blocks, i.e. subexpression protected from variation,
expression nodes may be tagged with \code{buildingBlockTags}. TODO
}

