\name{defaultGPFunctionAndConstantSets}
\alias{arithmeticFunctionSet}
\title{Default function- and constant factory sets for Genetic Programming...}
\description{Default function- and constant factory sets for Genetic Programming}
\details{\code{arithmeticFunctionSet}: \code{arithmeticFunctionSet} is an untyped function set containing the functions
"+", "-", "*", and "/".
\code{expLogFunctionSet} is an untyped function set containing the functions
"sqrt", "exp", and "ln".
\code{trigonometricFunctionSet} is an untyped function set containing the functions
"sin", "cos", and "tan".
\code{mathFunctionSet} is an untyped function set containing all the above functions.

\code{numericConstantSet} is an untyped constant factory set containing a single
constant factory that creates numeric constants via calls to \code{runif(1, -1, 1)}.

}
\alias{expLogFunctionSet}
\alias{trigonometricFunctionSet}
\alias{mathFunctionSet}
\alias{numericConstantSet}

