\name{expressionComposing}
\alias{subexpressions}
\title{Functions for decomposing and recombining R expressions...}
\usage{subexpressions(expr)
}
\description{Functions for decomposing and recombining R expressions}
\details{\code{subExpressions} returns a list of all subexpressions (subtrees) of an
expression \code{expr}.}
\arguments{\item{expr}{An R expression.}
}

