\name{exprLabel}
\alias{exprLabel}
\title{Return the "label" at the Root Node of an Expression Tree...}
\usage{exprLabel(expr)
}
\description{Return the "label" at the Root Node of an Expression Tree}
\details{Internal tool function that returns the function name if \code{expr} is a call,
or otherwise just \code{expr} itself.}
\value{The expression's root label.}
\arguments{\item{expr}{The expression to return the root label for.}
}

