\name{individualAnalysis}
\alias{inputVariablesOfIndividual}
\title{Functions for analysing GP individuals...}
\usage{inputVariablesOfIndividual(ind, inset)
}
\description{Functions for analysing GP individuals}
\details{\code{inputVariablesOfIndividual} returns a list of input variables in \code{inset} that are
used by the GP individual \code{ind}.}
\arguments{\item{ind}{A GP individual, represented as a R function.}
\item{inset}{A set of input variables.}
}

