\name{popfitness}
\alias{popfitness}
\title{Calculate the fitness value of each individual in a population...}
\usage{popfitness(pop, fitnessfunc)
}
\description{Calculate the fitness value of each individual in a population}
\value{A list of fitness function values in the same order as \code{pop}.}
\arguments{\item{pop}{A population of functions.}
\item{fitnessfunc}{The fitness function.}
}

