% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rangeTypeOfType}
\alias{rangeTypeOfType}
\title{Return the range type if t is a function type, otherwise just return t}
\usage{
rangeTypeOfType(t)
}
\arguments{
\item{t}{The type to extract the range type from.}
}
\value{
The range type.
}
\description{
Return the range type if t is a function type, otherwise just return t
}

