\name{rangeTypeOfType}
\alias{rangeTypeOfType}
\title{Return the range type if t is a function type, otherwise just return t...}
\usage{rangeTypeOfType(t)
}
\description{Return the range type if t is a function type, otherwise just return t}
\value{The range type.}
\arguments{\item{t}{The type to extract the range type from.}
}

