\docType{package}
\name{rgp-package}
\alias{rgp-package}
\title{The RGP package}
\description{
RGP is a simple yet flexible modular Genetic Programming
system for the R environment. The system implements
classical untyped tree-based genetic programming as well as
more advanced variants including, for example, strongly
typed genetic programming and Pareto genetic programming.
}
\author{
Oliver Flasch \email{oliver.flasch@fh-koeln.de}, Olaf
Mersmann \email{olafm@statistik.tu-dortmund.de}, Thomas
Bartz-Beielstein
\email{thomas.bartz-beielstein@fh-koeln.de}, Martin
Zaefferer \email{martin.zaefferer@fh-koeln.de}, Joerg Stork
\email{joerg.stork@fh-koeln.de}
}
\keyword{package}

