% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smse}
\alias{smse}
\title{Scaled mean squared error (SMSE)}
\usage{
smse(x, y)
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list.}
}
\value{
The NMSE between \code{x} and \code{y}.
}
\description{
Calculates the MSE between vectors after scaling them.
Beware that this error measure is invariant to scaling with negative constants, i.e.
the multiplicative inverse of the true functions also receives an error of 0.
See \url{http://www2.cs.uidaho.edu/~cs472_572/f11/scaledsymbolicRegression.pdf} for
details.
}

