\name{sortBy}
\alias{sortBy}
\title{Sort a vector or list by the result of applying a function...}
\usage{sortBy(xs, byFunc)
}
\description{Sort a vector or list by the result of applying a function}
\details{Sorts a vector or a list by the numerical result of applying the function \code{byFunc}.}
\value{The result of sorting \code{xs} by \code{byfunc}.}
\arguments{\item{xs}{A vector or list.}
\item{byFunc}{A function from elements of \code{xs} to \code{numeric}.}
}

