\name{sortByRange}
\alias{sortByRange}
\title{Tabulate a list of functions or input variables by the range part of their sTypes...}
\usage{sortByRange(x)
}
\description{Tabulate a list of functions or input variables by the range part of their sTypes}
\value{A table of the objects keyed by their range sTypes.}
\arguments{\item{x}{A list of functions or input variables to sort by range sType.}
}

