% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortByRange}
\alias{sortByRange}
\title{Tabulate a list of functions or input variables by the range part of their sTypes}
\usage{
sortByRange(x)
}
\arguments{
\item{x}{A list of functions or input variables to sort by range sType.}
}
\value{
A table of the objects keyed by their range sTypes.
}
\description{
Tabulate a list of functions or input variables by the range part of their sTypes
}

