\name{sortByRanking}
\alias{sortByRanking}
\title{Sort a vector or list via a given ranking...}
\usage{sortByRanking(xs, ranking=rank(xs))
}
\description{Sort a vector or list via a given ranking}
\details{Reorders a vector or list according to a given ranking \code{ranking}.}
\value{The result of reordering \code{xs} by \code{ranking}.}
\arguments{\item{xs}{The vector or list to reorder.}
\item{ranking}{The ranking to sort \code{xs} by, defaults to \code{rank(xs)}.}
}

