% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortByType}
\alias{sortByType}
\title{Tabulate a list of functions or input variables by their sTypes}
\usage{
sortByType(x)
}
\arguments{
\item{x}{A list of functions or input variables to sort by sType.}
}
\value{
A table of the objects keyed by their sTypes.
}
\description{
Tabulate a list of functions or input variables by their sTypes
}

